feat: create school vertical slice

+++ backend/rate-my-driving-instructor-backend/src/main/java/com/codecool/ratemydrivinginstructorbackend/controller/dto/NewInstructorDTO.java
@@ -1,4 +1,4 @@
 package com.codecool.ratemydrivinginstructorbackend.controller.dto;
 
-public record NewInstructorDTO(String firstName, String lastName, int schoolId) {
+public record NewInstructorDTO(String firstName, String lastName, long schoolId) {
 }
